requirements.txt
praw==7.7.1
python-dotenv==1.0.0
moviepy==1.0.3
Pillow==10.2.0
requests==2.31.0
python-slugify==8.0.1
loguru==0.7.2
gTTS==2.5.1
transformers==4.38.2
torch==2.2.1
sentencepiece==0.2.0
accelerate==0.27.2
pyyaml==6.0
numpy<2